# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Garage SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: staffLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffLoginResult"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getStaffMe
      tags: [auth]
      responses:
        "200":
          description: Current staff user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
        "401":
          description: Unauthorized

  # ── PORTAL (Customer) ─────────────────────────────────────────────────────────
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"

    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

    put:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /customers/{id}/vehicles:
    get:
      operationId: getCustomerVehicles
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
  /customers/{id}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /customers/{id}/balance:
    get:
      operationId: getCustomerBalance
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer balance and debt info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerBalance"

  # ── VEHICLES ──────────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceList"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/payments:
    get:
      operationId: getInvoicePayments
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payments for invoice
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: addPayment
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── TECHNICIANS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StaffLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        remember_me:
          type: boolean
    StaffUser:
      type: object
      required: [id, email, role, fullName]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        garageId:
          type: ["integer", "null"]
    StaffLoginResult:
      type: object
      required: [user, redirect]
      properties:
        user:
          $ref: "#/components/schemas/StaffUser"
        redirect:
          type: string
...
    Customer:
      type: object
      required: [id, fullName, phonePrimary]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phonePrimary:
          type: string
        phoneSecondary:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        ownerType:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        vehiclesCount:
          type: integer
        totalOrders:
          type: integer
        totalSpent:
          type: number
    CustomerList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
    CustomerInput:
      type: object
      required: [fullName, phonePrimary]
      properties:
        fullName:
          type: string
        phonePrimary:
          type: string
        phoneSecondary:
          type: string
        email:
          type: string
        address:
          type: string
        gender:
          type: string
        ownerType:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        fullName:
          type: string
        phonePrimary:
          type: string
        phoneSecondary:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
    CustomerBalance:
      type: object
      required: [totalInvoiced, totalPaid, balanceDue]
      properties:
        totalInvoiced:
          type: number
        totalPaid:
          type: number
        balanceDue:
          type: number
        overdueInvoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"

    # ── VEHICLE ────────────────────────────────────────────────────────────────
    Vehicle:
      type: object
      required: [id, customerId, plateNumber, make, model, year]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        plateNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: ["string", "null"]
        bodyType:
          type: string
        fuelType:
          type: string
        transmission:
          type: string
        currentMileage:
          type: ["integer", "null"]
        engineSize:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Order:
      type: object
      required: [id, orderNumber, vehicleId, customerId, status]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        vehicleId:
          type: integer
        customerId:
          type: integer
        status:
          type: string
        priority:
          type: string
        customerName:
          type: string
        vehiclePlate:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        estimatedCostMin:
          type: ["number", "null"]
        estimatedCostMax:
          type: ["number", "null"]
        grandTotal:
          type: ["number", "null"]
        receivedAt:
          type: string
        estimatedDeliveryAt:
          type: ["string", "null"]
        actualDeliveryAt:
          type: ["string", "null"]
        assignedTechnicianId:
          type: ["integer", "null"]
        assignedTechnicianName:
          type: ["string", "null"]
        paymentStatus:
          type: string
        customerComplaint:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Invoice:
      type: object
      required: [id, invoiceNumber, orderId, total, status]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        vehicleInfo:
          type: string
        subtotal:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        totalPaid:
          type: number
        balanceDue:
          type: number
        currency:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        issuedAt:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        notesAr:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
    Payment:
      type: object
      required: [id, invoiceId, amount, paymentMethod, paymentDate]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
        paymentReference:
          type: ["string", "null"]
        paymentDate:
          type: string
        receivedByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [amount, paymentMethod, paymentDate]
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        paymentReference:
          type: string
        paymentDate:
          type: string
        notes:
          type: string

    # ── TECHNICIAN ─────────────────────────────────────────────────────────────
...