# lib/api-zod/src/generated/types/staffLoginResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Garage SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffUser } from './staffUser';

export interface StaffLoginResult {
  user: StaffUser;
  redirect: string;
}