# lib/api-zod/src/generated/types/staffUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Garage SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffUser {
  id: number;
  email: string;
  role: string;
  fullName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  garageId?: number | null;
}